replace checks from getRequestParams.ts
